{"schema":"decodex/commit/1","summary":"Split status worktree projection","authority":"XY-1115"}
